backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.18.3"
fastapi = "^0.82.0"
python-multipart = "^0.0.5"
email-validator = "^1.2.1"
requests = "^2.28.1"
celery = "^5.2.7"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
tenacity = "^8.0.1"
emails = "^0.6"
sentry-sdk = "^1.9.8"
gunicorn = "^20.1.0"
jinja2 = "^3.1.2"
psycopg2-binary = "^2.9.3"
alembic = "^1.8.1"
sqlalchemy = "^1.4.41"
pytest = "^7.1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = "^1.10.2"

[tool.poetry.dev-dependencies]
mypy = "^0.971"
black = "^22.8.0"
isort = "^5.10.1"
autoflake = "^1.5.3"
flake8 = "^5.0.4"
pytest = "^7.1.3"
sqlalchemy-stubs = "^0.4"
pytest-cov = "^3.0.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

